requirements.txt
PyQt5==5.15.4
ursina==4.1.1
numpy>=1.19.0
pyserial==3.5
vpython==7.6.2
pyqtgraph==0.12.3